refactor(suite): split Spark SDK helpers
