+++ suite-common/spark/src/sdk/getErrorMessage.ts
@@ -1,0 +1,11 @@
+export const getErrorMessage = (error: unknown) => {
+    if (error instanceof Error) {
+        return error.message;
+    }
+
+    if (typeof error === 'string') {
+        return error;
+    }
+
+    return 'Unknown Spark SDK error';
+};
