+++ suite-common/spark/src/sdk/getTransferDirection.ts
@@ -1,0 +1,6 @@
+import { type SparkTransfer } from '../feature/sparkFeatureReducer';
+
+type SparkTransferRecord = Record<string, unknown>;
+
+export const getTransferDirection = (transfer: SparkTransferRecord): SparkTransfer['direction'] =>
+    transfer.transferDirection === 'OUTGOING' ? 'send' : 'receive';
